Refactor pack types endpoints into PackTypeController
