+++ src/controller/AdminController.java
@@ -27,7 +27,6 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 import model.admin.FormField;
-import model.bloodbagtype.BloodBagType;
 import model.compatibility.CrossmatchType;
 import model.donationtype.DonationType;
 import model.tips.Tips;
@@ -44,7 +43,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import repository.BloodBagTypeRepository;
 import repository.CrossmatchTypeRepository;
 import repository.DonationTypeRepository;
 import repository.FormFieldRepository;
@@ -55,7 +53,6 @@ import repository.TipsRepository;
 import repository.UserRepository;
 import repository.WorksheetTypeRepository;
 import utils.PermissionConstants;
-import viewmodel.PackTypeViewModel;
 
 @RestController
 public class AdminController {
@@ -70,9 +67,6 @@ public class AdminController {
 
   @Autowired
   LocationRepository locationRepository;
-
-  @Autowired
-  BloodBagTypeRepository bloodBagTypesRepository;
 
   @Autowired
   DonationTypeRepository donationTypesRepository;
@@ -271,14 +265,6 @@ public class AdminController {
     return map;
   }
 
-  @RequestMapping(value="/packtypes", method=RequestMethod.GET)
-  public  Map<String, Object> configureBloodBagTypesFormGenerator() {
-
-    Map<String, Object> map = new HashMap<String, Object>();
-    addAllBloodBagTypesToModel(map);
-    return map;
-  }
-
   @RequestMapping(value="/backupdata", method=RequestMethod.GET)
   @PreAuthorize("hasRole('"+PermissionConstants.MANAGE_BACKUP_DATA+"')")
   public void backupData(
@@ -346,9 +332,7 @@ public class AdminController {
     m.put("allDonationTypes", donationTypesRepository.getAllDonationTypes());
   }
 
-  private void addAllBloodBagTypesToModel(Map<String, Object> m) {
-    m.put("allBloodBagTypes", getPackTypeViewModels(bloodBagTypesRepository.getAllBloodBagTypes()));
-  }
+
 
   private void addAllCrossmatchTypesToModel(Map<String, Object> m) {
     m.put("allCrossmatchTypes", crossmatchTypesRepository.getAllCrossmatchTypes());
@@ -417,32 +401,6 @@ public class AdminController {
     addAllCrossmatchTypesToModel(map);
     return map;
   }
-  
-   @RequestMapping(value = "/packtypes/{id}", method = RequestMethod.GET)
-    @PreAuthorize("hasRole('" + PermissionConstants.MANAGE_BLOOD_BAG_TYPES + "')")
-    public ResponseEntity<BloodBagType> getPackTypeById(@PathVariable Integer id){
-        Map<String, Object> map = new HashMap<String, Object>();
-        BloodBagType packType = bloodBagTypesRepository.getBloodBagTypeById(id);
-        map.put("packtype", new PackTypeViewModel(packType));
-        return new ResponseEntity(map, HttpStatus.OK);
-    }
-
-    @RequestMapping(value = "/packtypes", method = RequestMethod.POST)
-    @PreAuthorize("hasRole('" + PermissionConstants.MANAGE_BLOOD_BAG_TYPES + "')")
-    public ResponseEntity savePackType(@Valid @RequestBody BloodBagType packType){
-        packType = bloodBagTypesRepository.saveBloodBagType(packType);
-        return new ResponseEntity(new PackTypeViewModel(packType), HttpStatus.CREATED);
-    }
-  
-    @RequestMapping(value = "/packtypes/{id}", method = RequestMethod.PUT)
-    @PreAuthorize("hasRole('" + PermissionConstants.MANAGE_BLOOD_BAG_TYPES + "')")
-    public ResponseEntity updateBloodBagType(@RequestBody BloodBagType packType , @PathVariable Integer id){
-        Map<String, Object> map = new HashMap<String, Object>();
-        packType.setId(id);
-        packType = bloodBagTypesRepository.updateBloodBagType(packType);
-        map.put("packtype", new PackTypeViewModel(packType));
-        return new ResponseEntity(map, HttpStatus.OK);
-    }
 
   @RequestMapping(value = "/donationtypes/{id}", method = RequestMethod.GET)
   @PreAuthorize("hasRole('"+PermissionConstants.MANAGE_DONATION_TYPES+"')")
@@ -505,15 +463,6 @@ public class AdminController {
         return new ArrayList<InetAddress>();
     }
     return listOfServerAddresses;
-  }
-  
-  private List<PackTypeViewModel> getPackTypeViewModels(List<BloodBagType> packTypes){
-      
-      List<PackTypeViewModel> viewModels = new ArrayList<PackTypeViewModel>();
-      for(BloodBagType packtType : packTypes){
-          viewModels.add(new PackTypeViewModel(packtType));
-      }
-      return viewModels;
   }
 }
